+++ phinet/__init__.py
@@ -1,0 +1,2 @@
+from .networks import PhiNet
+from .blocks import PhiNetConvBlock
